Text.Microstache

Description

This is a Haskell implementation of Mustache templates. The implementation conforms to the version 1.1.3 of official Mustache specification https://github.com/mustache/spec. It is extremely simple and straightforward to use with minimal but complete API — three functions to compile templates (from directory, from file, and from lazy text) and one to render them.

The implementation uses the Megaparsec parsing library to parse the templates which results in superior quality of error messages.

For rendering you only need to create Aeson's Value where you put the data to interpolate. Since the library re-uses Aeson's instances and most data types in Haskell ecosystem are instances of classes like ToJSON, the whole process is very simple for end user.

Template Haskell helpers for compilation of templates at compile time are available in the Text.Microstache.Compile.TH module. The helpers are currently available only for GHC 8 users though.

One feature that is not currently supported is lambdas. The feature is marked as optional in the spec and can be emulated via processing of parsed template representation. The decision to drop lambdas is intentional, for the sake of simplicity and better integration with Aeson.

Here is an example of basic usage:

{-# LANGUAGE OverloadedStrings #-}

module Main (main) where

import Data.Aeson
import Data.Text
import Text.Megaparsec
import Text.Microstache
import qualified Data.Text.Lazy.IO as TIO

main :: IO ()
main = do
  let res = compileMustacheText "foo"
        "Hi, {{name}}! You have:\n{{#things}}\n  * {{.}}\n{{/things}}\n"
  case res of
    Left err -> putStrLn (parseErrorPretty err)
    Right template -> TIO.putStr $ renderMustache template $ object
      [ "name"   .= ("John" :: Text)
      , "things" .= ["pen" :: Text, "candle", "egg"]
      ]

If I run the program, it prints the following:

Hi, John! You have:
  * pen
  * candle
  * egg

For more information about Mustache templates the following links may be helpful:

The official Mustache site: https://mustache.github.io/

The manual: https://mustache.github.io/mustache.5.html

The specification: https://github.com/mustache/spec

Stack Builders Stache tutorial: https://www.stackbuilders.com/tutorials/haskell/mustache-templates/

Types

data Template

Mustache template as name of “top-level” template and a collection of all available templates (partials).

Template is a Semigroup. This means that you can combine Templates (and their caches) using the (<>) operator, the resulting Template will have the same currently selected template as the left one. Union of caches is also left-biased.

Constructors

Template

Fields

templateActual :: PName

Name of currently “selected” template (top-level one).

templateCache :: Map PName [Node]

Collection of all templates that are available for interpolation (as partials). The top-level one is also contained here and the “focus” can be switched easily by modifying templateActual.

data Node

Structural element of template.

Constructors

TextBlock Text
Plain text contained between tags

EscapedVar Key
HTML-escaped variable

UnescapedVar Key
Unescaped variable

Section Key [Node]
Mustache section

InvertedSection Key [Node]
Inverted section

Partial PName (Maybe Word)
Partial with indentation level (Nothing means it was inlined)

newtype Key

Identifier for values to interpolate.

The representation is the following:

[] — empty list means implicit iterators;

[text] — single key is a normal identifier;

[text1, text2] — multiple keys represent dotted names.

Constructors

Key

Fields

unKey :: [Text]

newtype PName

Identifier for partials. Note that with the OverloadedStrings extension you can use just string literals to create values of this type.

Constructors

PName

Fields

unPName :: Text

data MustacheException

Exception that is thrown when parsing of a template has failed or referenced values were not provided.

Constructors

MustacheParserException ParseError
Template parser has failed. This contains the parse error.

MustacheRenderException PName Key

Deprecated: Not thrown anymore, will be removed in the next major version of microstache

A referenced value was not provided. The exception provides info about partial in which the issue happened PName and name of the missing key Key.

displayMustacheException :: MustacheException -> String

Since: 1.0.1

data MustacheWarning

Since: 1.0.1

Constructors

MustacheVariableNotFound Key
The template contained a variable for which there was no data counterpart in the current context

MustacheDirectlyRenderedValue Key
A complex value such as an Object or Array was directly rendered into the template

displayMustacheWarning :: MustacheWarning -> String

Since: 1.0.1

Compiling

compileMustacheDir

Arguments

:: PName
Which template to select after compiling

-> FilePath
Directory with templates

-> IO Template
The resulting template

Compile all templates in specified directory and select one. Template files should have extension mustache, (e.g. foo.mustache) to be recognized. This function does not scan the directory recursively.

The action can throw the same exceptions as getDirectoryContents, and readFile.

compileMustacheFile

Arguments

:: FilePath
Location of the file

-> IO Template

Compile single Mustache template and select it.

The action can throw the same exceptions as readFile.

compileMustacheText

Arguments

:: PName
How to name the template?

-> Text
The template to compile

-> Either ParseError Template
The result

Compile Mustache template from a lazy Text value. The cache will contain only this template named according to given PName.

Rendering

renderMustache :: Template -> Value -> Text

Render a Mustache Template using Aeson's Value to get actual values for interpolation.

renderMustacheW :: Template -> Value -> ([MustacheWarning], Text)

Like renderMustache but also return a list of warnings.

Since: 1.0.1