Text.Microstache.Parser

Description

Megaparsec parser for Mustache templates. You don't usually need to import the module, because Text.Microstache re-exports everything you may need, import that module instead.

Documentation

parseMustache

Arguments

:: FilePath
Location of file to parse

-> Text
File contents (Mustache template)

-> Either ParseError [Node]
Parsed nodes or parse error

Parse given Mustache template.