Text.Microstache.Render

Description

Functions for rendering Mustache templates. You don't usually need to import the module, because Text.Microstache re-exports everything you may need, import that module instead.

Documentation

renderMustache :: Template -> Value -> Text

Render a Mustache Template using Aeson's Value to get actual values for interpolation.

renderMustacheW :: Template -> Value -> ([MustacheWarning], Text)

Like renderMustache but also return a list of warnings.

Since: 1.0.1