Text.Microstache.Type

Description

Types used by the package. You don't usually need to import the module, because Text.Microstache re-exports everything you may need, import that module instead.

Documentation

data Template

Mustache template as name of “top-level” template and a collection of all available templates (partials).

Template is a Semigroup. This means that you can combine Templates (and their caches) using the (<>) operator, the resulting Template will have the same currently selected template as the left one. Union of caches is also left-biased.

Constructors

Template

Fields

templateActual :: PName

Name of currently “selected” template (top-level one).

templateCache :: Map PName [Node]

Collection of all templates that are available for interpolation (as partials). The top-level one is also contained here and the “focus” can be switched easily by modifying templateActual.

data Node

Structural element of template.

Constructors

TextBlock Text
Plain text contained between tags

EscapedVar Key
HTML-escaped variable

UnescapedVar Key
Unescaped variable

Section Key [Node]
Mustache section

InvertedSection Key [Node]
Inverted section

Partial PName (Maybe Word)
Partial with indentation level (Nothing means it was inlined)

newtype Key

Identifier for values to interpolate.

The representation is the following:

[] — empty list means implicit iterators;

[text] — single key is a normal identifier;

[text1, text2] — multiple keys represent dotted names.

Constructors

Key

Fields

unKey :: [Text]

showKey :: Key -> Text

Pretty-print a key, this is helpful, for example, if you want to display an error message.

newtype PName

Identifier for partials. Note that with the OverloadedStrings extension you can use just string literals to create values of this type.

Constructors

PName

Fields

unPName :: Text

data MustacheException

Exception that is thrown when parsing of a template has failed or referenced values were not provided.

Constructors

MustacheParserException ParseError
Template parser has failed. This contains the parse error.

MustacheRenderException PName Key

Deprecated: Not thrown anymore, will be removed in the next major version of microstache

A referenced value was not provided. The exception provides info about partial in which the issue happened PName and name of the missing key Key.

displayMustacheException :: MustacheException -> String

Since: 1.0.1

data MustacheWarning

Since: 1.0.1

Constructors

MustacheVariableNotFound Key
The template contained a variable for which there was no data counterpart in the current context

MustacheDirectlyRenderedValue Key
A complex value such as an Object or Array was directly rendered into the template

displayMustacheWarning :: MustacheWarning -> String

Since: 1.0.1